reverted more changes to return to original commit

+++ crypto/packet.c
@@ -13,7 +13,7 @@
 
 #define DEFAULT_BUF_SIZE    256
 
-int WPACKET_allocate_bytes(WPACKET *pkt, size_t len, unsigned char **allocbytes)
+int WPACKET_allocate_bytes(WPACKET *pkt, unsigned int len, unsigned char **allocbytes)
 {
     if (!WPACKET_reserve_bytes(pkt, len, allocbytes))
         return 0;
@@ -379,7 +379,7 @@ int WPACKET_start_sub_packet(WPACKET *pkt)
     return WPACKET_start_sub_packet_len__(pkt, 0);
 }
 
-int WPACKET_put_bytes__(WPACKET *pkt, size_t val, size_t size)
+int WPACKET_put_bytes__(WPACKET *pkt, unsigned int val, size_t size)
 {
     unsigned char *data;
